Rework design for video manager
Video players will register themselves to get their handles. They will use this handle to interact with the video manager, and by extension, the relays. The video manager will carry much of the heavy lifting of retrying videos, swapping out video players, switching to the next video, and making sure videos are loaded. It is still up to the video player however to ensure that the videos are in sync across the network.