Media Controller now inherits from Callback Receiver
This removes a large chunk of duplicate code I've been wanting to get rid of for a while. I've created an override and made the base send callback method virtual, though it was meant to hide the base member. This should be resolved with a future version of U#. For now, this method is not intended to be virtual, so please do not override it. You will run into trouble in the future if you do.